backend/ec2/live-gemini/pyproject.toml
[project]
name = "live-gemini"
version = "0.1.0"
description = ""
authors = [
    {name = "steven-passynkov",email = "passynkovsteven@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "google-genai (>=1.21.1,<2.0.0)",
    "boto3 (>=1.38.41,<2.0.0)",
    "vertexai (>=1.71.1,<2.0.0)",
    "opensearch-py (>=3.0.0,<4.0.0)",
    "requests-aws4auth (>=1.3.1,<2.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "deepgram-sdk (>=4.3.1,<5.0.0)",
    "langgraph (>=0.4.8,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
